requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
transformers
accelerate
torch
pickle-mixin
